+++ myuw_vue/components/_common/collapsed-notice.vue
